Syntax error is now handled
I added an array of flags where each flag corresponds to an index in the gateList. To make it consist in length as gate list grows in size it is passed along other parameters to the MenuButtons.
In case of any syntax error that may raise interrupt erros in the backend, we instead send a pop up window message saying"fix the syntax" and the state of the forntend circuit is sat back to the last correct gateList.